+++ src/hocs/__tests__/renderInRouter.test.tsx
@@ -1,7 +1,7 @@
 import { screen, fireEvent } from '@testing-library/react'
 
-import { renderInRouter } from '../../../index'
-import { EXAMPLE_HOME_PAGE_ROUTE_NAME, ExampleAppRoutes } from '../../../routes'
+import { renderInRouter } from '../../index'
+import { EXAMPLE_HOME_PAGE_ROUTE_NAME, ExampleAppRoutes } from '../../routes'
 
 describe('renderInRouter HOC', () => {
   it('Renders component', () => {
